traefik: enable access logs
This will write the access logs to stdout, so they will end up
in the docker logging system.

This logs the request target, the response status, the IP, and the user agent.
Similar to what nginx logs by default, but for all our containers, and in json.

Also bumps the log size for traefik to 1GB, so we keep one day worth of logs.