rsync: disable traefik timeout and set a rsyncd timeout instead
Starting with the v2.11.2 release traefik added a default read timeout
of 60 seconds. This resulted in 7 of our mirrors failing to rsync
completely.

Bumping this to 300 didn't seem to help (though I did only wait 10 hours)
so this disables it completely (we are down to 2 failing with this rigth now).

Let's just disable this completely and instead set the timeout option in
the rsync config. Hoping that rsync knows better how to decide when a connection
is of no use anymore. Use 600 because it's recommended in the rsyncd.conf man page.