repo,mirrorbits: add HTTP redirects for old mingw/(i686,x86_64) directories
Many years ago we switched to a the directory structure matching the repo names,
to simplify the pacman repo config, see
https://github.com/msys2/MSYS2-packages/commit/cfdb662d9859a21174afa3c9c3d1eb9ec1d98727
For backwards compat we added symlinks for the old names back then, so the old paths
would continue to work.

The duplication there has lead to confusion though, as it's not clear why some repos
are duplicated, and if there is a difference between them.

Quite some time has passed now, but according to the logs there are still some
(outdated) systems accessing the old paths.

Since it's not that hard, add a HTTP redirect for them. This means, once we remove
the symlinks, those paths will stop working on the mirrors but continue to work
on the main repo at least.